@@ -4,7 +4,7 @@ import * as firebase from './firebase'
 import * as localstorage from './localstorage'
 
 const drivers: StorageDriverMap = { firebase, localstorage }
-const chosenDriver = process.env.VUE_APP_STORAGE_DRIVER || 'localstorage'
+const chosenDriver = import.meta.env.VITE_STORAGE_DRIVER || 'localstorage'
 const driver = drivers[chosenDriver]
 
 if (!driver) {
